Split DynamicQuestion into components by q.type
