+++ FeedBackApp.Frontend/src/components/feedback/questions/types.ts
@@ -1,0 +1,9 @@
+import { Question } from "@/models/StudentContext";
+
+export type BaseQuestionProps = {
+    q: Question;
+    index: number;
+    value: string | string[];
+    onChange: (val: string | string[]) => void;
+    isInvalid?: boolean;
+};
